bugs found, added options, and buttons
Found a bug when dealing with placing cards in the foundation. when i clicked an 8 with a 7 on top, it took both to the foundation and the seven was on the wrong pile. the new game button just refreshes the page, but it asks the user first. the quit button just alerts at the moment. and the undo button prints to the console. i think the next step should be fixing the bug, and then do the undo button.